+++ lib/providers/notifiers/planted_notifier.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:seeds/models/models.dart';
 import 'package:seeds/providers/services/http_service.dart';
+import 'package:seeds/v2/datasource/remote/model/planted_model.dart';
 
 class PlantedNotifier extends ChangeNotifier {
   PlantedModel balance;
